refactor: move code part generating single multiversion card to the separate file
